Ty Package Tracker for 1.5.x
This module adds the ability for your customers and admin to track packages from different carriers. By default, the carriers are FedEx, UPS, and USPS. You can easily add or change the carriers. To add a new carrier, the URL tracking link is required.

The official support thread can be found at:
https://www.zen-cart.com/showthread.php?36126-Track-Package-from-FedEx-UPS-Postal-DHL-etc