Send shipping info to the UPS WorldShip software
This module will send shipping information to the UPS WorldShip software to create shipping labels. You must have UPS Worldship software installed in order for this module to work.

This is the final beta version I complete in 2008 but it was never uploaded to the plugin section. I am releasing this to the public so that others can use it and update it if need. I no longer use UPS so I can not offer any support for this module. However, it has been tested and a few beta tester from 2008 are still using it.

Official support thread for this module can be at: https://www.zen-cart.com/showthread.php?213496-Ty-UPS-WorldShip-Support-Thread